fix: map page
